requirements.txt
python-dotenv
pydantic
aiogram
rcon
mcstatus
pandas
psutil
prometheus_client
matplotlib
plotly
sqlalchemy
pytest
schedule
apscheduler
aiohttp
Jinja2
